pyproject.toml
[project]
name = "pe-unde-bei-apa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "flask>=3.1.3",
    "openai>=2.30.0",
    "pymupdf>=1.27.2.2",
    "pymysql[rsa]>=1.1.2",
    "requests>=2.33.0",
    "rowordnet>=1.1.0",
    "sentence-transformers>=5.3.0",
    "setuptools==69.5.1",
    "spacy>=3.8.14",
    "torch>=2.11.0",
    "yoyo-migrations>=9.0.0",
]
